[DmaCell] Prototype implementation for a DmaCell
This adds a DmaCell to more explicitly capture the mutability of buffers
when Dma is occuring. This serves to hold a buffer undergoing Dma until
the user notifies the DmaCell, using an unsafe api, that the Dma has
completed. At this time, the buffer is returned to the user.

Note, we are not considering the length field set for Dma here. If this
is mistakenly set, we can read/write past the buffer.